Fix publish script.

+++ Utilities/Publish.ps1
@@ -141,7 +141,9 @@ Get-ChildItem -Path "$Root\ScreenCast.Win\bin\x64\Release\" -Exclude "*.xml" | C
 $PublishDir = "$Root\Desktop.Win\bin\Release\app.publish"
 &"$MSBuildPath" "$Root\Desktop.Win" /t:Publish /p:Configuration=Release /p:Platform=AnyCPU
 # Copy to download folder.
-Get-ChildItem -Path "$Root\Server\wwwroot\Downloads\WinDesktop" | Remove-Item -Force -Recurse
+if (Test-Path -Path "$Root\Server\wwwroot\Downloads\WinDesktop") {
+    Get-ChildItem -Path "$Root\Server\wwwroot\Downloads\WinDesktop" | Remove-Item -Force -Recurse
+}
 Get-ChildItem -Path "$PublishDir" | Move-Item -Destination "$Root\Server\wwwroot\Downloads\WinDesktop" -Force
 Rename-Item -Path "$Root\Server\wwwroot\Downloads\WinDesktop\setup.exe" -NewName "Remotely_Setup.exe" -Force
